Collect distinct user_ids in observed_users subquery
The MAX_OBSERVED_USERS constant is a restriction on the number of users
that we return from the observed_users helper, not on the number of
enrollments that we look at when determining who an observer is
observing. By collecting distinct associated_user_ids in the subquery,
we'll collect up to 1000 users and not be limited by enrollments with
duplicate associated_user_ids.

closes LS-3208
flag = none

Test plan:
 - Login as an observer who's observing several students
 - Expect that all the observed students still appear in the dropdown

Change-Id: I023966476298b7f1bc186d80a246b05ab1cd498b
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/293749
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Ed Schiebel <eschiebel@instructure.com>