(webpack) unify tinymce installation
refs FOO-2835
flag = none

hoist tinymce to the top since we're using the same version as the RCE,
this way we don't end up with 2 separate installations

that strange CSS require() call has to be relative for a reason i don't
have neither time nor desire to unravel

there was a discrepancy in the noParse directive between our webpack
config in test and outside of test: the test one was excluding the
entirety of tinymce, which is fine for the main tinymce module but not
so for the "icons" one that the rce package depends on. This was working
OK because the test config was ignoring the top-level tinymce icons
module, which is never required, but now that it's getting required,
it fell apart. I kept to the non-test version where we exclude only the
main module "tinymce.js", as there's no value beyond that.

~ test plan ~

the RCE still works (e.g. in assignment edit page) and looks OK

Change-Id: I664930753189d8f3966ddec0d0340ba100af6ab3
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/294132
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Charley Kline <ckline@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>