spec: Skip flaky specs for plugins
Refs FOO-2994
flag=none

Something happened recently to cause the specs in this file to
fail in the post-merge build. It is likely timing-related and
thus fixable, but the spec probably also needs general improvement
in terms of what it's trying to test, and we may also decide to
simply remove the test since what is is testing is only available
to siteadmin anyway and thus there is low risk from breakage.

Test plan:
* Merge it and see if the post-build merge passes?

Change-Id: I31d4f02815cf3f915bfd37dd9114446d424ae9cd
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/295169
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob Burroughs <jburroughs@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>