include current account and sub-accounts for user role filter search
fixes FOO-3008
flag = none

Test plan:

Pre-requisites: An account with one course and a student enrolled.
A sub-account under the first account with a course and student
enrolled in the course.

Go to the people page in the parent account and note that both
students are listed without any filter applied

Apply the Student filter and note that both students enrolled
in a course in the parent account and sub-account are listed

Change-Id: I494a033dee8aa39d73c81e3d2dc6a86b07163f10
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/295923
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob Burroughs <jburroughs@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>