Save warning before navigating away from SpeedGrader
Added modal warning whenever teacher edits rubric in SpeedGrader and
attempts to leave the page without saving the rubric.

flag=none
closes EVAL-2325

Test Plan:
1. Have a course with at least one student and a teacher/admin
2. Create an assignment with a rubric attached to it
and one criteria on the rubric
3. As a teacher/admin, go into the SpeedGrader
for the assignment that was created
4. Click the View Rubric button and modify a grade
but do not press the save button
5. Try to navigate away from the page (such as pressing the
assignment name link on the left of the top bar) and you should
see a warning modal.
6. Go back to the SpeedGrader and press View Rubric again.
7. Do not modify a grade and do not press the save button.
8. Try to navigate away from the page again
and you should not see a warning modal
9. Go back to the SpeedGrader and press View Rubric again.
10. Modify a grade and press the save button
11. Press the View Rubric button and modify a grade
12. Try to navigate away from the page
and you should see a warning modal.

Change-Id: If390ae8a524d979a49fcc7a459800ecbcf6dd0d4
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/295982
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Aaron Shafovaloff <ashafovaloff@instructure.com>
Product-Review: Jody Sailor
Reviewed-by: Eduardo Escobar <eduardo.escobar@instructure.com>
Reviewed-by: Aaron Shafovaloff <ashafovaloff@instructure.com>