Revert "Remove inheritance from conditional_release"
This reverts commit 0e5262aa842812aa9a0064604214baf4af0a3808.

Inheritance was the right way to go. It was how the feature flag worked
and this brings it back to working as that did. One thing to note about
settings is that inheritance works for a default value but once a value
is set it will use that. For example, if the account has Mastery Paths
enabled and unlocked then new courses will have the setting turned on.
If the course turns it off then it will no longer use the inheritance so
changing the account value without locking it won't affect the courses
that have a saved setting value unless the lock is used.

fixes LS-3288
flag=none

test plan:
  - Have an account
  - Create a sub account and a course in the sub account
  - Turn on Mastery Path but leave unlocked on the account
  - Verify the subaccount has Mastery Path on but unlocked
  - Verify the course has Mastery Path on
  - Turn off Mastery Paths on the subaccount but leave unlocked
  - Verify the course has Mastery Paths off but can turn it on
  - Lock Mastery Paths on the subaccount but leave it off
  - Verify the course has Mastery Paths off and can't turn it on
  - Lock Mastery Paths on the account and turn it on
  - Verify the subaccount and course have mastery paths on and locked

Change-Id: I4b849a75f97a2a9aeceff91049e6c26871ca0963
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/296571
Reviewed-by: Luis Oliveira <luis.oliveira@instructure.com>
QA-Review: Luis Oliveira <luis.oliveira@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Allison Howell <allison.howell@instructure.com>