Fix potential contrast issue with CondensedButtons
Fixes FOO-2981
flag=none

The use of InstUI <CondensedButton> to replace the
deprecated <Button variant="link"> created some issues
with customers who decided to invert the contrast of
their "Primary Button" style in the theme editor. To
wit, if the Primary Button is changed from something
light-on-dark to something dark-on-light, then an
instance of a CondensedButton would have very poor
contrast, as its text is rendered in the color of the
background of a regular Button.

Not sure what to do about this issue with Condensed-
Button but this commit replaces those instances with
<Link as="button"> and explicitly styled <Text>
which results in something that looks nearly identical
and behaves the same in a screen reader.

Test plan:
* See the Steps to Reproduce in the ticket
* Visit accounts/self and accounts/self/users and
  notice the column headings on the lists; they should
  still work properly and also still be readable even
  though you changed the color of "Primary Button" to
  white
* Also check accounts/self/permissions and the permission
  names running down the left side of the table should
  be similarly readable

Change-Id: I2cb60abc3010ed0aac5513363a02222a24c87a89
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/298407
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>