Fix calendar_events API for students and hidden/deleted accounts
Students don't have the :read permission for root accounts which
causes some issues when students try to set custom colors to the
account calendars, and also makes the api/calendar_events fails when
a new account is added, it happens only the first time cause the
second time the user tries to request the API the context is already
in the user_preferences.

This PS adds an extra validation to the GET api/v1/calendar_events
API to use the same criteria we have in api/v1/account_calendars and
let the account be visible to any user that has the account
associated

closes LS-3391, LS-3392
flag= account_calendar_events

Test plan:

Setup:
Enable account_calendar_events feature flag
Create an account
Create a course in the account you created and enroll a student
Go to the account calendars settings page
Enable the account you created
Go to the Calendar, add the account to the other calendars section
using the modal
Create an event in the account.

LS-3392
Log in as the student and go to the calendar
Add the account calendar using the modal
Expect to see the account calendar in the sidebar and the event in
the calendar
Expect not to see 401 errors

LS-3391
Log in as admin
Go to the account calendar settings and hide the account
Go back to the calendar; notice that the calendar does not show the
account calendar anymore, and the event you created is no longer in
the calendar
Go to settings and enable the account calendar again
Expect to see everything related to that account in the calendar
again
Delete the course and the account.
Go back to the calendar and expect not to see either the account or
the events in the calendar.

Change-Id: Ieb25bbccc94bf40a55bf33e482f36b9bd9f49200
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/299669
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jackson Howe <jackson.howe@instructure.com>
QA-Review: Jackson Howe <jackson.howe@instructure.com>
Product-Review: Jonathan Guardado <jonathan.guardado@instructure.com>