Add user's AssignmentSetActions to modules cache key
Modules erb templates are cached extensively using action controller
caching (note that this is distinct from the regular rails caching or
request caching and is not enabled by default in Canvas development
mode). This cache entry has no TTL, so it is only cleared when the key
is updated or Redis discards the entry. Under certain conditions, a
user can find themself with a stale entry, resulting in them seeing
the "Locked until mastery path processes..." notice in modules. This
can occur because the notice is displayed according to the user's
ConditionalRelease::AssignmentSetAction entries for the module, but
these entries are not considered in the cache key. This commit
hashes those entries' IDs and includes the hash in the cache key.

closes LS-3508
refs LS-3488
flag = none

Test plan:
 - Set syncronous: false in config/initializers/delayed_job.rb:239
   to force jobs to run asyncronously in Canvas development
 - Start your Canvas server with
   `ACTION_CONTROLLER_CACHING=true rails s` to enable action
   controller caching in Canvas development
 - Create a course, create a module, enable mastery paths in the
   course, publish the course, enroll a student
 - In the module, create a classic quiz and an assignment
 - Add an essay question to the quiz
 - Set the quiz mastery paths so that the bottom (third) scoring
   range assigns the assignment; the other scoring ranges have no
   assignments
 - Edit the assignment to assign it only to 'Mastery Paths'
 - Submit the quiz as the student
 - Stop your delayed jobs process to simulate a slow/borked job
 - Grade the quiz submission; ensure the grade does NOT fall in the
   bottom mastery paths scoring range
 - As the student, visit modules
 - Expect to see "Locked until mastery path is processed..."
 - Restart jobs process
 - Expect the "Locked..." message to be gone and for the assignment
   to not be visible to the student
   (previously, the locked message would have persisted)

Change-Id: I6000487e27452f5e392631382d3bf3d640e8dffb
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/302422
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Luis Oliveira <luis.oliveira@instructure.com>
QA-Review: Luis Oliveira <luis.oliveira@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>