Don't link to homeroom announcement for students
Make the announcement title a link for teachers, or for students in
a subject. Don't show the link for students on the dashboard - they're
not intended to see the homeroom course internals and there's nothing
about the announcement in the homeroom course that the student can't
see from the dashboard anyway.

closes LS-3460
flag = none

Test plan:
 - Create an announcement in a homeroom course and in a subject
 - Check out the announcement title as a teacher on the dashboard;
   should be a link
 - Check out the announcement title as a teacher in a subject; should
   be a link
 - Check out the announcement title as a student on the dashboard;
   should NOT be a link
 - Check out the announcement title as a student in a subject;
   should be a link

Change-Id: I11c7e8e33a5a143c867f3d08ec292c9e9221c983
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/302739
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>