Update to cope with the RCE in native browser fullscreen (#126)
* Update to cope with the RCE in native browser fullscreen

The primary change is to checker.js where it forwards the mountNode
prop to the Tray. This gets the a11y checker on top of the RCE
while in fullscreen.
Adjacent to that we position the indicator so it's correct and
visible whether in fullscreen or not.

Also fixed a bug where adjust() kept tracking indicators even
after they were removed from the DOM, and throttled the work
adjust does to just once every 10ms.

Moved the es build output from lib/modules to es and added
jest.config.js to ignore anything in either directory.

Updated the yarn "test" script so you can now execute
yarn test path/to/test/file.js
and jest will execute just the given file.

Added a yarn "test:debug" to facilitate debugging specs.

Updated the version number and changelog in preparation for
publishing the updates.

closes MAT-1115

* Update to cope with the RCE in native browser fullscreen

The primary change is to checker.js where it forwards the mountNode
prop to the Tray. This gets the a11y checker on top of the RCE
while in fullscreen.
Adjacent to that we position the indicator so it's correct and
visible whether in fullscreen or not.

Also fixed a bug where adjust() kept tracking indicators even
after they were removed from the DOM, and throttled the work
adjust does to just once every 10ms.

Moved the es build output from lib/modules to es and added
jest.config.js to ignore anything in either directory.

Updated the yarn "test" script so you can now execute
yarn test path/to/test/file.js
and jest will execute just the given file.

Added a yarn "test:debug" to facilitate debugging specs.

Updated the version number and changelog in preparation for
publishing the updates.

closes MAT-1115

* Update to cope with the RCE in native browser fullscreen

The primary change is to checker.js where it forwards the mountNode
prop to the Tray. This gets the a11y checker on top of the RCE
while in fullscreen.
Adjacent to that we position the indicator so it's correct and
visible whether in fullscreen or not.

Also fixed a bug where adjust() kept tracking indicators even
after they were removed from the DOM, and throttled the work
adjust does to just once every 10ms.

Moved the es build output from lib/modules to es and added
jest.config.js to ignore anything in either directory.

Updated the yarn "test" script so you can now execute
yarn test path/to/test/file.js
and jest will execute just the given file.

Added a yarn "test:debug" to facilitate debugging specs.

Updated the version number and changelog in preparation for
publishing the updates.

closes MAT-1115