Reuse AssignmentOverrideStudent objects
For this issue there may be a race condition being presented that is hard
to recreate. When the assignment override student objects are deleted
they may not be persisting to the database before the new objects are
created. This patchset introduces a new approach to reuse an existing
assignment override student object instead of deleting and creating new
ones. This is beneficial to stop the race condition by querying for an
object instead of deleting and creating. It also lowers the amount of
extraneous data being saved since the assignment override students are
not actually deleted but instead have their workflow_state set to deleted.

fixes LS-3591
flag=course_pacing

test plan:
- Create a course pace with students
- Change the dates on the pace and republish
- Verify the job succeeded and all dates are present

Change-Id: I3154f1b711c600fd1d5eadeec153ceb7f424506d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/306739
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Luis Oliveira <luis.oliveira@instructure.com>
QA-Review: Luis Oliveira <luis.oliveira@instructure.com>
Product-Review: Eric Saupe <eric.saupe@instructure.com>