remove assignment visibility for concluded student
Before this change, if an assignment was assigned to a section that a
student was concluded in AND that student had another enrollment in
the course that was active, the student would maintain visibility of
any assignments that had a due date in the future. This meant that the
students could submit for that assignment. After this change,
students will no longer be able to see assignments that are assigned
to sections that concluded/completed in that have a due date in the
future. The will continue to see any assignments that were due before
their enrollment was completed.

closes EVAL-2748
flag=none

Test Plan 1:
* Pre-requisites: Migrations have been run, and a student enrolled in
two sections of the same course (A & B).

1. As the teacher, create one assignment in section A.
2. As the teacher, create two assignments in section B.
   - create one assignment with a due date in the past
   - create one assignment with a due date in the future
3. As the student, view your assignments. User should see all
   assignments
4. As teacher, go to the user page for the
   student (courses/:course_id/users/:id), click 'more user details...',
   and then click on 'Conclude this Enrollment' for Section B
5. As the student, view your assignments. User should now see:
   - Assignment for Section A (still enrolled)
   - Assignment for Section B with due date in past
6. Verify the student cannot see the assignment for Section B that
   has the due date in the future.

Change-Id: I68a5ae908077d2f591532513a022e339db2c8c3a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/306831
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
Reviewed-by: Kai Bjorkman <kbjorkman@instructure.com>
Reviewed-by: Cameron Ray <cameron.ray@instructure.com>
QA-Review: Spencer Olson <solson@instructure.com>
QA-Review: Kai Bjorkman <kbjorkman@instructure.com>
Product-Review: Deborah Kwak <deborah.kwak@instructure.com>
Migration-Review: Jacob Burroughs <jburroughs@instructure.com>