Prevent RCE selection from being cleared while not in focus
When a user selects something or places their cursor somewhere
in the RCE, then blurs the RCE, then invoked the browser Find
command, their selection or cursor position in the RCE is cleared,
and if they then go on to insert content, it will be placed
in the wrong location. This solves the issue by storing the selection
on blur and restoring it on focus and before commands are executed.

Closes MAT-1138

flag=none

Test plan:
- Set up an RCE with several lines of content
- Select one of the lines (or just place cursor on one of them)
- Blur the RCE by going to insert Course Link or similar
- Invoke the browser find command and search for a junk string
- Click on one of the course links or similar to insert it
- Ensure that the link was added to the correct place in the content
- Repeat above for several types of content inserting (images, docs,
  links, etc...)

Change-Id: I410a780e9d35b357b29ad5cd41dbf73985414e69
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/306870
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Mysti Lilla <mysti@instructure.com>
Reviewed-by: Jake Oeding <jake.oeding@instructure.com>
Product-Review: David Lyons <lyons@instructure.com>
QA-Review: Mysti Lilla <mysti@instructure.com>