Add session token to LTI 1.1 sessionless launches
This should help with NQ issues where the tool launches in mobile but it
doesn't have access to Canvas resources such as images.
generate_session_token requires the endpoint be hit with an API token,
but I think this should be the case for the mobile requests (I don't
know if anyone uses this endpoint without an API token)

As mentioned in the comment, we _might_ be able to simplify LTI 1.1
sessionless launches to just make a session token and then do a normal
launch, like we do for LTI 1.3 tools. I left some additional info in
a JIRA comment.

closes INTEROP-7912
flag=none

Test plan:
- Create a sessionless launch for an LTI 1.1 tool by hitting the API at
  /api/v1/accounts/self/external_tools/sessionless_launch?id=123
where 123 is the id of your LTI 1.1 tool.
- In a new incognito window, visit the URL returned
- The tool should launch. Then in the same incognito visit Canvas -- you
  should still be logged in.

Change-Id: I7fff6e6e558a93af6659007475e16017951fa50d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/309488
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>
Reviewed-by: Tucker Mcknight <tmcknight@instructure.com>
QA-Review: Tucker Mcknight <tmcknight@instructure.com>