Make redux store work for all of the rce
and not just the CanvasContentTray
I considered making CanvasContentTray and all its descendents
a redux connected component, but that would have been a massive
change both in how state values were passed to the components
and for review. I elected to simply pass the store's state
to CanvasContentTray as a prop, just like when it was rendered
from witin the <StoreProvider>. Then none of the descendents
needed to change.

How the store persisted which accordion section was opened last
wasn't simply changed by the action that opened the new panel,
but by a subscription to store changes. Now that the store state
persists between CanvasContentTray openings, we can simply
save it to session storage in the reducer.

I also converted RCEWrapper.test.js to jest
which made me update the tinymce mock
and a tinymce Editor mock so you can actually
mount an RCEWrapper now and have it work (within reason)

The specs for canvas_mentions would not pass using the new
FakeEditor and after spending time trying to update them
I gave up and moved the old mock to live along side the
only tests that use it.

closes MAT-1195
flag=none

test plan:
  - with devTools open to the network panel
  - open the CanvasContentTray (course documents etc.)
  - close the tray and wait an hour
  - open the tray
  > expect to see a call to refresh the jwt
  - close the tray
  - open the tray
  > expect no call to refresh the jwt

  - open 'course links'
  - expand one of the accordion sections
  - close the tray
  - open 'course links'
  > expect the same section to be open

Change-Id: I31904a133d760138d6b9bd85a6ab569620454ba4
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/310508
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Yona Appletree <yona.appletree@instructure.com>
QA-Review: Yona Appletree <yona.appletree@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>